masterfinancer@v10

+++ src/main/java/com/elegro/masterfinan/infraestructura/cruds/AbonoDaoRepository.java
@@ -3,5 +3,5 @@ package com.elegro.masterfinan.infraestructura.cruds;
 import com.elegro.masterfinan.infraestructura.dao.DaoRecordLong;
 import com.elegro.masterfinan.infraestructura.entity.Abono;
 
-public interface AbonoDaoRepository extends DaoRecordLong<Abono, Integer> {
+public interface AbonoDaoRepository extends DaoRecordLong<Abono, Long> {
 }
